fractional number of packs Part 01

+++ server/repository/migrations/postgres/2021-09-15T10-00_create_stock_line_table/up.sql
@@ -7,8 +7,8 @@ CREATE TABLE stock_line (
     expiry_date DATE,
     cost_price_per_pack DOUBLE PRECISION NOT NULL,
     sell_price_per_pack DOUBLE PRECISION NOT NULL,
-    available_number_of_packs INTEGER NOT NULL,
-    total_number_of_packs INTEGER NOT NULL,
+    available_number_of_packs DOUBLE PRECISION NOT NULL,
+    total_number_of_packs DOUBLE PRECISION NOT NULL,
     pack_size INTEGER NOT NULL,
     on_hold BOOLEAN NOT NULL,
     note TEXT
